feat: Logo, Search 아이콘 추가, GNB 내 NavigationDiv 분리
- Logo, Search SVG 파일 및 아이콘 컴포넌트 추가
- GNB 내부 NavigationDiv 컴포넌트 분리
