+++ assets/icons/index.ts
@@ -2,3 +2,5 @@ export { default as NaverIcon } from './naver.svg';
 export { default as KakaoIcon } from './kakao.svg';
 export { default as GoogleIcon } from './google.svg';
 export { default as FacebookIcon } from './facebook.svg';
+export { default as Logo } from './logo.svg';
+export { default as Search } from './search.svg';
